+++ substrafl/remote/substratools_methods.py
@@ -88,8 +88,11 @@ class RemoteMethod:
         elif OutputIdentifiers.shared in outputs:
             self.save_model(method_output, outputs[OutputIdentifiers.shared])
 
-        elif OutputIdentifiers.performance in outputs:
-            tools.save_performance(method_output, outputs[OutputIdentifiers.performance])
+        else:
+            for output_id in outputs:
+                # The performances are the only identifier user defined.
+                if output_id not in list(OutputIdentifiers):
+                    tools.save_performance(method_output[output_id], outputs[output_id])
 
     def generic_function(
         self,